+++ examples/send-transfer-with-priority/app.ts
@@ -15,7 +15,6 @@ export const start = async () => {
           from: '[ETH_ACCOUNT]',
           to: '0x6009608a02a7a15fd6689d6dad560c44e9ab61ff',
           value: '100000000',
-          transactionType: 'dynamic_fee',
           gasPricePolicy: {
             priority: Priority.VeryHigh
           }