+++ examples/send-transfer/app.ts
@@ -14,8 +14,7 @@ export const start = async () => {
         params: {
           from: '[ETH_ACCOUNT]',
           to: '0x6009608a02a7a15fd6689d6dad560c44e9ab61ff',
-          value: '100000000',
-          transactionType: 'legacy'
+          value: '100000000'
         }
       },
       'ExampleTransfer'